Update FRC1633 Java2018v022318
This is now adding the camera. It is easier since it is the USB camera we do not have to do much in order to actually get to work. Note this example is directly from the WPI lib sooooo idk if it will work 100% nor what some of those variables really do. Another thing to note down is that there are is a way to connect to the RoboRio here is the link:

https://wpilib.screenstepslive.com/s/currentCS/m/vision/l/288981-using-the-microsoft-lifecam-hd-3000

This gives us the instructions. It is also included in the code. Next Autonomous...