Update FRC1633 Java2018v022318
Oke soooooo in this branch I have added the auto code for it. I created another branch so if there are any changes that need to be made to the auto it can be done without any trouble of altering the already code things.

The encoders are next along with the rest of what is need form Mentor B's email. 

I am also creating a RobotMap class where like for FTC I can map everything but that still in the fog since all the things that go into the mapping part and naming is a little tricky. That will still require me to further add more branches if not repositories.

-Geo